src/txts_unit/hz_client/requirements.txt
click==8.1.3
hazelcast-python-client==5.1
importlib-metadata==4.11.3
typer==0.4.1
typing_extensions==4.2.0
zipp==3.8.0
Twisted==22.10.0
python-dotenv==0.21.0
netifaces==0.11.0
